Create implementation PrelegentAdapter.
